[Bronze V] Title: 꼬마 정민, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Bronze/11382. 꼬마 정민/꼬마 정민.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(void) {
-    long long num1,num2,num3;
-    scanf("%lld %lld %lld",&num1,&num2,&num3);
+    long num1,num2,num3;
+    scanf("%ld %ld %ld",&num1,&num2,&num3);
     
-    printf("%lld\n",num1+num2+num3);
+    printf("%ld\n",num1+num2+num3);
     return 0;
 }
